requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
gunicorn==23.0.0
whitenoise==6.8.2
python-decouple==3.8
dj-database-url==2.2.0
